+++ src/Starcounter/Query/Execution/Enumerators/ReferenceLookup.cs
@@ -247,7 +247,7 @@ internal class ReferenceLookup : ExecutionEnumerator, IExecutionEnumerator
     /// <summary>
     /// Saves the underlying enumerator state.
     /// </summary>
-    public unsafe Int32 SaveEnumerator(Byte* keysData, Int32 globalOffset, Boolean saveDynamicDataOnly)
+    public unsafe UInt16 SaveEnumerator(Byte* keysData, UInt16 globalOffset, Boolean saveDynamicDataOnly)
     {
         return globalOffset;
     }
